Fleshed things out; basics on TM & TT
A basic home page, linking to Tux Paint (external site, maintained
by yours truly), and Markdown pages for Tux, of Math Command,
and Tux Typing.

Linking to old SourceForge.net file sections, since they seem to
be as up-to-date as I can tell (based on looking at the old
Tux4Kids.alioth.debian.org site), and no actual files seem to be
hosted in the GitHub projects.  Also, of course, linking to the
GitHub project pages, since that seems to be the new place the
source is maintained.